+++ CHANGELOG.md
@@ -82,6 +82,7 @@ FUTURE PLANS
 ## [0.19.8] - 2025-05-10
 ### Fixed
 - z index of character image is too high and interferes with touchscreen playing
+- NaN showing for notehistory when localstorage has no previously stored value
 
 ## [0.19.7] - 2025-05-04
 ### Added